Python/Django/simple_encrypt/simpleproject/requirements.txt
Django==2.1.7
pycrypto==2.6.1
pytz==2018.9
